pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "canadian-bookkeeping-agent"
version = "0.1.0"
description = "Multi-agent audit system for Canadian bookkeepers' QuickBooks Online work, delivered via Microsoft Teams."
requires-python = ">=3.11"
dependencies = [
    "botbuilder-core>=4.16.0",
    "botbuilder-schema>=4.16.0",
    "botbuilder-integration-aiohttp>=4.16.0",
    "aiohttp>=3.9.0",
    "anthropic>=0.40.0",
    "pdfplumber>=0.11.0",
    "openpyxl>=3.1.0",
    "pandas>=2.2.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.6.0",
]
ocr = [
    "pytesseract>=0.3.10",
    "pdf2image>=1.17.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.setuptools.package-data]
"src.bot" = ["cards/*.json"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
